// packages/libera-ai-web/src/@types/errors.d.ts
export type ApiErrorMessage =
  | 'INVALID_CREDENTIALS'
  | 'FORBIDDEN_RESOURCE'
  | 'TOKEN_NOT_PROVIDED'
  | 'INVALID_TOKEN'
  | 'USER_NOT_FOUND'
  | 'USER_ALREADY_ACTIVATED'
  | 'EMAIL_ALREADY_EXISTS'
  | 'CPF_ALREADY_EXISTS'
  | 'EMAIL_OR_PASSWORD_INVALID'
  | 'WEEK_PASSWORD'
  | 'UNAUTHORIZED_LOGIN'